+++ src/Symfony/Component/DependencyInjection/Loader/XmlFileLoader.php
@@ -501,6 +501,7 @@ class XmlFileLoader extends FileLoader
                         throw new InvalidArgumentException(sprintf('Tag "<%s>" with type="service_locator" only accepts maps of type="service" references in "%s".', $name, $file));
                     }
                     break;
+                case 'tagged':
                 case 'tagged_iterator':
                 case 'tagged_locator':
                     $type = $arg->getAttribute('type');