+++ src/Symfony/Component/DomCrawler/Crawler.php
@@ -571,7 +571,7 @@ class Crawler implements \Countable, \IteratorAggregate
     /**
      * Returns the text of the first node of the list.
      *
-     * Pass true as the 2nd argument to normalize whitespaces.
+     * Pass true as the second argument to normalize whitespaces.
      *
      * @param string|null $default             When not null: the value to return when the current node is empty
      * @param bool        $normalizeWhitespace Whether whitespaces should be trimmed and normalized to single spaces
@@ -580,7 +580,7 @@ class Crawler implements \Countable, \IteratorAggregate
      *
      * @throws \InvalidArgumentException When current node is empty
      */
-    public function text(string $default = null, bool $normalizeWhitespace = false)
+    public function text(string $default = null, bool $normalizeWhitespace = true)
     {
         if (!$this->nodes) {
             if (null !== $default) {
@@ -592,7 +592,7 @@ class Crawler implements \Countable, \IteratorAggregate
 
         $text = $this->getNode(0)->nodeValue;
 
-        if (\func_num_args() > 1 && func_get_arg(1)) {
+        if ($normalizeWhitespace) {
             return trim(preg_replace('/(?:\s{2,}+|[^\S ])/', ' ', $text));
         }