+++ src/Symfony/Component/Lock/CHANGELOG.md
@@ -14,7 +14,10 @@ CHANGELOG
  * added InvalidTtlException  
  * deprecated `StoreInterface` in favor of `BlockingStoreInterface` and `PersistingStoreInterface`
  * `Factory` is deprecated, use `LockFactory` instead
-
+ * `StoreFactory::createStore` allows PDO and Zookeeper DSN. 
+ * deprecated services `lock.store.flock`, `lock.store.semaphore`, `lock.store.memcached.abstract` and `lock.store.redis.abstract`, 
+   use `StoreFactory::createStore` instead.
+    
 4.2.0
 -----